pyproject.toml
[project]
name = "streamlit-deep-agents"
version = "0.1.0"
description = "Streamlit UI for LangChain's Deep Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.0.3",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.30",
    "streamlit>=1.48.1",
    "tavily-python>=0.7.10",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "langgraph-cli[inmem]>=0.3.6",
]
